A new folder and new file have been added. Some old files have been deleted. A couple of files have been fixed.
Foulder name has changed.

Changed the Query call.

Added a method.

File has been fixed.
