+++ commands/delete_semester.rb
@@ -1,6 +1,7 @@
-require_relative 'database'
+require_relative '../database'
+require_relative '../queries/semester_query'
 
-module Command
+module Commands
   class DeleteSemester
     def initialize
       @db = Database.new
@@ -8,9 +9,7 @@ module Command
 
     def call
       input_name
-
       delete_semester
-
       puts 'Семестр был успешно удален.' if @result.ntuples.positive?
     ensure
       @db.close
@@ -22,7 +21,6 @@ module Command
       loop do
         puts 'Введите название Семестра, который хотите удалить:'
         @semester_name = gets.chomp
-
         break if valid_name?
       end
     end
@@ -30,8 +28,7 @@ module Command
     def valid_name?
       return puts 'Название Семестра не может быть пустым' if @semester_name.nil? || @semester_name.strip.empty?
 
-      @semester_id = FindSemesterByName.new(db: @db, semester_name: @semester_name).call
-
+      @semester_id = Queries::SemesterQuery.new.find_by_name(semester_name: @semester_name)
       return puts "Семестр с именем #{@semester_name} не найден." if @semester_id.nil?
 
       true
