+++ queries/base_query.rb
@@ -1,0 +1,7 @@
+require_relative '../database'
+
+class BaseQuery
+  def initialize
+    @db = Database.new
+  end
+end
