+++ semester.rb
@@ -1,5 +1,3 @@
-require_relative 'list_semester'
-
 class Semester
   attr_accessor :semester_name, :start_date, :end_date
 
@@ -7,9 +5,5 @@ class Semester
     @semester_name = semester_name
     @start_date = start_date
     @end_date = end_date
-  end
-
-  def self.list_semesters
-    ListSemester.perform
   end
 end
